+++ convert_to_hf.py
@@ -1,13 +1,30 @@
 import json
 import os
 import re
+import shutil
 
 import torch
 from tqdm.auto import trange
 
-from src.saving import add_inference_code, update_config
-from src.utils import pack_int_data
 from transformers import AutoConfig, PretrainedConfig
+
+
+def get_int_dtype(nbits: int) -> torch.dtype:
+    if nbits <= 8:
+        return torch.int8
+    if nbits <= 16:
+        return torch.int16
+    if nbits <= 32:
+        return torch.int32
+    if nbits <= 64:
+        return torch.int64
+    raise ValueError(f"No dtype available for {nbits}-bit codebooks")
+
+
+@torch.inference_mode()
+def pack_int_data(data: torch.IntTensor, nbits: int) -> torch.IntTensor:
+    data[data >= 2 ** (nbits - 1)] -= 2**nbits
+    return data.to(get_int_dtype(nbits))
 
 
 def get_num_layers(config) -> int:
@@ -58,6 +75,53 @@ def get_metadata(in_path: os.PathLike) -> dict:
     }
 
 
+def update_config(old_config: PretrainedConfig, aqlm_metadata: dict[str, int]):
+    old_config_type = type(old_config)
+    old_model_type = old_config.model_type
+    new_model_type = f"{old_model_type}_aqlm"
+
+    class AqlmConfig(old_config_type):
+        model_type = new_model_type
+
+        def __init__(
+            self,
+            aqlm: dict[str, int] = {
+                "nbits_per_codebook": 16,
+                "num_codebooks": 1,
+                "out_group_size": 8,
+                "in_group_size": 1,
+            },
+            **kwargs,
+        ):
+            super().__init__(**kwargs)
+            self.aqlm = aqlm
+
+    config_dict = old_config.to_dict()
+    config_dict["auto_map"] = {
+        "AutoConfig": f"configuration_{new_model_type}.{old_config.__class__.__name__}",
+        "AutoModelForCausalLM": f"modeling_{new_model_type}.{config_dict['architectures'][0]}",
+    }
+    del config_dict["_name_or_path"]
+
+    new_config = AqlmConfig(
+        {
+            "nbits_per_codebook": aqlm_metadata["nbits_per_codebook"],
+            "num_codebooks": aqlm_metadata["num_codebooks"],
+            "out_group_size": aqlm_metadata["out_group_size"],
+            "in_group_size": aqlm_metadata["in_group_size"],
+        }
+    )
+    new_config.update(config_dict)
+    return new_config
+
+
+def add_inference_code(model_type: str, save_path: os.PathLike):
+    if os.path.isdir(f"./transformers/{model_type}"):
+        shutil.copytree(f"./transformers/{model_type}", save_path, dirs_exist_ok=True)
+    else:
+        print(f"No predefined PreTrainedModel exists for {model_type}. You'll have to copy-paste some code yourself.")
+
+
 if __name__ == "__main__":
     import argparse
 
@@ -81,7 +145,6 @@ if __name__ == "__main__":
     args = parser.parse_args()
 
     old_config = AutoConfig.from_pretrained(args.model)
-    print(type(old_config))
     metadata = get_metadata(args.in_path)
 
     add_inference_code(old_config.model_type, args.out_path)
